Add @CanIgnoreReturnValue to some methods in MatchAssert
This is an AssertJ annotation, but IDEA knows about it (and ErrorProne
probably too). It prevents a warning in the IDE that the return value is
ignored, which is a false positive in this case, because we don't have
to string more assertions. This makes the use optional without
suppressing any more warnings inside the method.