Support pattern matching in area reduced to window frame
In pattern matching with the MATCH_RECOGNIZE clause,
the whole partition is available for the pattern search.
In pattern matching in window, the pattern is searched
for each row within the specified window frame.

This change introduces the distinction between:
- partitionStart, partitionEnd -- being the partition bounds
and
- searchStart, searchEnd -- being the frame bounds.
It enables pattern search in reduced area of the frame.