Fix planning of pattern navigation offset
The offset value is a LongLiteral. Before this change,
it was rewritten by the TranslationMap, which could lead to
replacing the literal with a symbol reference, if the map
already contained a mapping for that literal.
It resulted in a failure in LogicalIndexExtractor, where
a literal was expected.

After this change, the offset argument is not rewritten.