Introduce hidden $approx_set aggregation
The semantics of function is the same as user visible approx_set, yet
$approx_set is defined for all types available in Trino, while
approx_set works only for BIGINT, VARCHAR(x) and DOUBLE.

The reason for intrducing new function and not changing approx_set is
that we did not want to introduce backward incompatible change for
approx_set.

It is still to be discussed if it making backward incompatible change to
approx_set is huge problem. If we decide it is not, we will rename
$approx_set to approx_set.