Accept that transaction is cancelled early in TestJdbcConnection
So query fails to start as transaction is already cancelled.
This is one of expected outcome (but rare) of query cancellation.

The one test that is not updated makes sure that user is able to read
data, so surely query has started there so `TRANSACTION_ALREADY_ABORTED`
is not expected there.