Test file access without JMX in Iceberg
This changes how `TrackingFileIoProvider` is installed in Iceberg
plugin and moves it to be a test-only code.

Previously, it was part of Iceberg plugin code (but inaccessible to
users). Due to this, tests accessed it via MBean invocations.

After the change, `TrackingFileIoProvider` is test-only entity,
instantiated directly in the test, and invoked "normally".