Add direct block construction helpers to BooleanType
Adds a contract method to BooleanType that indicates that callers
can legally choose to construct BooleanType blocks from byte[]
values directly. Adding this method for bypassing BlockBuilders
can be significant for BooleanType in particular, because the
per-row overhead of ByteArrayBlockBuilder#valueIsNull is the same
as for the values themselves, meaning each builder row is 2x larger
when no values can be null.

Also adds a similar utility to create single boolean value blocks
directly from BooleanType.