Fix side effect in decimal addition operator
addLongShortLong performs an in-place addition that mutates the value of
the Slice associated with "left". This is based on the assumption that
"rescale" always returns a new Slice, which would avoid modifications
to the input arguments to the operator. However, when the rescale factor
is 0, rescale just returns its input argument.