Remove transitionToRunning on Stage schedulingComplete()
The origin is `dead` code, since the check condition for each
task used this.getState() which is SqlstageExecution#getState().
Stage could nerver lead to RUNNING by this condition, actually
the line545 on updateTaskStatus() is the real effective way.

We remove it, because it is safe and can reduce the tasks iteration
checks and code is clear(there is only one point to RUNNING)