Fix planning of recursive queries with duplicate outputs of partial plan
If the recursion base relation or the recursion step relation
is planned so that it has duplicate output symbols, they need to
be re-mapped to distinct symbols. This is necessary to successfully
replace this part of the plan with another plan in the next recursion
step.

This is achieved by adding a disambiguating projection.