Move detailed storage formats tests from suite1 to suite2
Suite1 is at its execution time limits. Certain `TestHiveStorageFormats`
tests are not tagged as `storage_formats` because they do not need to be
executed on many environments, but this makes them included in Suite1.

This commits introduces new test group (`storage_formats_detailed`) and
assigns it to these tests, effectively moving them from Suite1 to
Suite2. It also reassigns `test*NullFormat` to the new group.