Fix Hive DELETE bug due to colliding split files
Though this bug was originally described as an "original
files bug", the root cause of the bug was a DELETE operation
on a bucket so large that it was partitioned into multiple
splits, all with the same statement number.  The bug was fixed
by passing a lambda to InternalHiveSplit that hands out
successive statementIds for each bucket number.