Don't abort the cursor when we're at the end of JDBC resultset
The JDBC `abort` operation is a heavy one - so heavy, in fact, that it
expects a separate `Executor` to offload it to. So we'd like to not call
it if we can avoid it.

The original rationale was that some JDBC drivers insist on draining the
resultset on `close` when there are still results to be returned, which
is a problem when we abort a query before it finishes execution. But we
don't have to do it unconditionally.

This commit short-circuits the abort call when the
`ResultSet#isAfterLast` returns `true`. This is an optional method and
it's not guaranteed that all al vendors implement it, though. So we push
this check to driver-specific implementations wherever it is necessary
and we know it's supported.