Create TaskStatus before stats when reporting taskInfo to coordinator
Task may have its stats populated and state updated to FINISHED during
the createTaskInfo() call, which could potentially create TaskInfo with
FINISHED state, but with some of the stats missing.
Creating TaskStatus first makes sure that stats are already present.

This handles a rare case of flaky tests reported in #5172